refactor(sample-app-routes): move the shared route patterns into a runtime-neutral package
