+++ apps/sample-app-routes/test/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  // The node --test graph; only here do node types exist.
+  "extends": "../../../tsconfig.base.json",
+  "compilerOptions": {
+    "noEmit": true,
+    "allowImportingTsExtensions": true,
+    "erasableSyntaxOnly": true,
+    "lib": ["ES2022"],
+    "types": ["node"]
+  },
+  "include": ["*.ts", "../src/**/*"]
+}
