+++ apps/sample-app-shared/package.json
@@ -11,7 +11,6 @@
     "format": "oxfmt \"**/*.{js,ts,json,md,css}\"",
     "format:check": "oxfmt --check \"**/*.{js,ts,json,md,css}\"",
     "lint": "oxlint .",
-    "test": "node --test \"src/**/*.test.ts\"",
     "typecheck": "tsc -p tsconfig.json --noEmit"
   },
   "dependencies": {
@@ -24,12 +23,12 @@
     "lit-dialog": "catalog:",
     "lit-ui-router": "workspace:*",
     "lodash": "catalog:",
+    "sample-app-routes": "workspace:*",
     "ui-router-navigation-location-plugin": "workspace:*"
   },
   "devDependencies": {
     "@types/dom-navigation": "catalog:",
     "@types/lodash": "catalog:",
-    "@types/node": "catalog:",
     "oxfmt": "catalog:",
     "typescript": "catalog:",
     "vite": "catalog:"
