+++ apps/sample-app-shared/src/app/contacts/states.ts
@@ -15,7 +15,7 @@ import { type Contact } from './interface.js';
 import EditContact from './EditContact.js';
 
 import { dsrRedirectToDefaultFromWithin } from '../util/dsr-default-redirect-within.js';
-import { routeSegments } from '../routes.js';
+import { routeSegments } from 'sample-app-routes';
 
 /**
  * This state displays the contact list.
