+++ docs/package.json
@@ -18,7 +18,7 @@
     "@spectrum-web-components/icons-workflow": "^1.12.1",
     "sample-app-lit-mobx": "workspace:*",
     "sample-app-lit-vanilla": "workspace:*",
-    "sample-app-shared": "workspace:*",
+    "sample-app-routes": "workspace:*",
     "screenfull": "^6.0.2",
     "vue": "^3.5.39"
   },
